ci: fix telescope

+++ database/migrations/2018_08_08_100000_create_telescope_entries_table.php
@@ -16,6 +16,11 @@ return new class extends Migration
 
     public function up(): void
     {
+        // Skip Telescope migrations during testing
+        if (!app()->environment('local')) {
+            return;
+        }
+
         $schema = Schema::connection($this->getConnection());
 
         $schema->create('telescope_entries', function (Blueprint $table) {
@@ -55,6 +60,11 @@ return new class extends Migration
 
     public function down(): void
     {
+        // Skip Telescope migrations during testing
+        if (!app()->environment('local')) {
+            return;
+        }
+
         $schema = Schema::connection($this->getConnection());
 
         $schema->dropIfExists('telescope_entries_tags');
